requirements.txt
fastapi[standard]==0.112.1
pydantic==2.8.2
uvicorn==0.30.6
python-dotenv==1.0.1
pymongo[srv]==4.9.1